launch Elasticsearch at login
Can't run two service calls because of https://github.com/chef/chef/issues/3285. So the symlink is created and the service is started after all of the plugins are installed. This guarantees that Elasticsearch and its plugins will be installed before attempting to start the service. But it also means that the service cannot be restarted by any other recipe or resource.

Resolves #2.